build: adopt task-input-aware Turbo planning
